Move declare cairo1 command test to separate file
