+++ tests/integration/commands/declare_cairo1_command_test/cairo_project.toml
