+++ tests/integration/commands/declare_cairo1_command_test/lib.cairo
